Refactor `WithURI` trait to `URL` helper
