+++ app/Helpers/URL.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Syntatis\FeatureFlipper\Concerns;
+namespace Syntatis\FeatureFlipper\Helpers;
 
 use function is_string;
 use function parse_url;
@@ -16,12 +16,16 @@ use const PHP_URL_PATH;
 /**
  * General methods to work with URIs and URLs.
  */
-trait WithURI
+class URL
 {
+	private function __construct()
+	{
+	}
+
 	/**
 	 * Retrieve the current request URL.
 	 */
-	private static function getCurrentURL(): string
+	public static function getCurrent(): string
 	{
 		$schema = isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on' ? 'https://' : 'http://';
 		$host = isset($_SERVER['HTTP_HOST']) && is_string($_SERVER['HTTP_HOST']) ? $_SERVER['HTTP_HOST'] : '';
@@ -33,7 +37,7 @@ trait WithURI
 	/**
 	 * Check if the current request is the WordPress login page.
 	 */
-	private static function isLoginURL(): bool
+	public static function isLogin(): bool
 	{
 		$urlLogin = wp_login_url();
 		$scriptName = isset($_SERVER['SCRIPT_NAME']) && is_string($_SERVER['SCRIPT_NAME']) ?
@@ -51,7 +55,7 @@ trait WithURI
 		}
 
 		// Try to identify if the login page is customized.
-		$url = self::getCurrentURL();
+		$url = self::getCurrent();
 		$urlPath = rtrim((string) parse_url($url, PHP_URL_PATH), '/');
 
 		return rtrim((string) parse_url($urlLogin, PHP_URL_PATH), '/') === $urlPath;
