+++ app/Modules/Security.php
@@ -9,9 +9,9 @@ use SSFV\Codex\Contracts\Hookable;
 use SSFV\Codex\Foundation\Hooks\Hook;
 use SSFV\Jaybizzle\CrawlerDetect\CrawlerDetect;
 use SSFV\Psr\Container\ContainerInterface;
-use Syntatis\FeatureFlipper\Concerns\WithURI;
 use Syntatis\FeatureFlipper\Features\LoginIdentifier;
 use Syntatis\FeatureFlipper\Helpers\Option;
+use Syntatis\FeatureFlipper\Helpers\URL;
 use WP_Error;
 
 use function define;
@@ -21,8 +21,6 @@ use const PHP_INT_MIN;
 
 class Security implements Hookable, Extendable
 {
-	use WithURI;
-
 	public function hook(Hook $hook): void
 	{
 		if (! Option::isOn('xmlrpc')) {
@@ -83,7 +81,7 @@ class Security implements Hookable, Extendable
 
 	public function blockBots(): void
 	{
-		if (! self::isLoginURL()) {
+		if (! URL::isLogin()) {
 			return;
 		}
